2010-11-12  Ramon Novoa  <rnovoa@artica.es>
	* windows/pandora_wmi.cc: isServiceRunning returns -1 for module
	  states other than "Running" and "Stopped".

	* modules/pandora_module_logevent.cc,
	  modules/pandora_module_logevent.h: Return a non empty string if
	  the event description can not be read. Save the event id as an
	  unsigned long int to avoid overflows.

	* modules/pandora_module_service.cc,
	  modules/pandora_module_service.h: Aesthetic fix. Removed carriage
	  returns.



git-svn-id: https://svn.code.sf.net/p/pandora/code/trunk@3580 c3f86ba8-e40f-0410-aaad-9ba5e7f4b01f

+++ pandora_agents/win32/modules/pandora_module_logevent.h
@@ -39,13 +39,13 @@ namespace Pandora_Modules {
 
 	class Pandora_Module_Logevent : public Pandora_Module {
 	private:
-        int id;
-		int type;
-		string source;
-		string application;
-		string pattern;
-		HANDLE log_event;
-		HANDLE messages_dll;
+        unsigned long id;
+	int type;
+	string source;
+	string application;
+	string pattern;
+	HANDLE log_event;
+	HANDLE messages_dll;
 
         HANDLE openLogEvent ();
         void closeLogEvent ();
